+++ src/MatrixSolver.cpp
@@ -315,7 +315,7 @@ void Lapack(const CalculationMatrix& A, const CalculationMatrix& B,const Calcula
 	Matrix1D<double> rhs(A.total_size);
 	for (size_t ix = 0; ix < psd.size_q1; ix++) {
 		for (size_t iy = 0; iy < psd.size_q2; iy++) {
-			int in = B.index1d(ix, iy);
+			int in = B.index1d(static_cast<int>(ix), static_cast<int>(iy));
 			rhs[in] = psd[ix][iy];
 		}
 	}
@@ -333,7 +333,7 @@ void Lapack(const CalculationMatrix& A, const CalculationMatrix& B,const Calcula
 	// copy back
 	for (size_t ix = 0; ix < psd.size_q1; ix++) {
 		for (size_t iy = 0; iy < psd.size_q2; iy++) {
-			int in = B.index1d(ix, iy);
+			int in = B.index1d(static_cast<int>(ix), static_cast<int>(iy));
 			psd[ix][iy] = rhs[in];
 		}
 	}
